requirements.txt
astroid==1.5.3
Babel==2.4.0
blinker==1.4
certifi==2017.4.17
chardet==3.0.4
click==6.7
coverage==4.4.1
decorator==4.0.11
defusedxml==0.5.0
docopt==0.6.2
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.5
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
ordereddict==1.1
pbr==3.1.1
pipreqs==0.4.7
pkg-resources==0.0.0
pylint==1.7.4
python3-openid==3.1.0
pytz==2017.2
rauth==0.7.0
requests==2.18.1
six==1.10.0
SQLAlchemy==1.1.11
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
urllib3==1.21.1
Werkzeug==0.12.2
Whoosh==2.7.4
wrapt==1.10.11
WTForms==2.1
yarg==0.1.9
